fix: add filter for active/inactive user in login

+++ Group-3_Warehouse-Management-System/src/main/java/com/example/controller/user/UserDetailServlet.java
@@ -31,7 +31,7 @@ public class UserDetailServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        User user = userService.findUserById(Integer.parseInt(request.getParameter("id")));
+        User user = userService.findUserById(Long.parseLong(request.getParameter("id")));
         List<Role> roles = roleDAO.findAll();
         request.setAttribute("roles", roles);
         request.setAttribute("user", user);
